requirements.txt
tabula-py==2.9.0
pandas==2.1.4
openpyxl==3.1.2